Move 'credentials' action buttons to DataGrid header
